requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database ORM
sqlalchemy>=2.0.0

# Data Validation
pydantic>=2.0.0

# HTTP Requests
requests>=2.31.0

# Web Scraping
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Environment Variables
python-dotenv>=1.0.0

# Date/Time Handling
python-dateutil>=2.8.0
